[#6] refactor: 로그인 페이지 리팩토링
